pyproject.toml
[project]
name = "rag-project"
version = "0.1.0"
description = "Medical Knowledge Graph RAG Chatbot"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Backend & API
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "python-dotenv>=1.2.2",
    "httpx>=0.27.0",
    # RAG & Vector Store
    "langchain>=0.1.0",
    "langchain-community>=0.1.0",
    "langchain-text-splitters>=0.1.1",
    "faiss-cpu>=1.8.0",
    "elasticsearch>=8.12.0",
    # Knowledge Graph (Neo4j & HGT)
    "neo4j>=5.18.0",
    "torch>=2.2.0",
    "transformers>=4.38.0",
    "torch-geometric>=2.5.0",
    # LLM & Cloud Services
    "groq>=0.5.0",
    "modal>=0.62.0",
    "google-generativeai>=0.4.0",
    # Database (Conversation History)
    "psycopg2-binary>=2.9.9",
    # Data Loading & Processing
    "pypdf>=4.1.0",
    "unstructured[pdf]>=0.12.0",
    "pandas>=2.2.0",
    # Evaluation
    "ragas>=0.1.5",
    "langchain-google-genai>=4.2.1",
]

[tool.uv]
managed = true
